+++ AzzyBot/Modules/Core/CoreAzzyStatsGeneral.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Reflection;
 using System.Threading.Tasks;
 using AzzyBot.ExceptionHandling;
+using AzzyBot.Modules.Core.Enums;
 using AzzyBot.Strings.Core;
 using Microsoft.Extensions.Logging;
 
@@ -97,10 +98,10 @@ internal static class CoreAzzyStatsGeneral
 
     internal static async Task<string> GetBotCommitAsync()
     {
-        if (CoreModule.CommitLock is null)
+        if (CoreModule.AzzyBotLock is null)
             return "Commit not found";
 
-        string commit = await CoreModule.CommitLock.GetFileContentAsync();
+        string commit = await CoreModule.AzzyBotLock.GetFileContentAsync(CoreFileValuesEnum.Commit);
         if (string.IsNullOrWhiteSpace(commit))
             commit = "Commit not found";
 
@@ -109,10 +110,10 @@ internal static class CoreAzzyStatsGeneral
 
     internal static async Task<string> GetBotCompileDateAsync()
     {
-        if (CoreModule.BuildTimeLock is null)
+        if (CoreModule.AzzyBotLock is null)
             return "CompileDate not found";
 
-        if (!DateTime.TryParse(await CoreModule.BuildTimeLock.GetFileContentAsync(), out DateTime dateTime))
+        if (!DateTime.TryParse(await CoreModule.AzzyBotLock.GetFileContentAsync(CoreFileValuesEnum.CompileDate), out DateTime dateTime))
             return "CompileDate not found";
 
         return $"<t:{CoreMisc.ConvertToUnixTime(dateTime)}>";