+++ AzzyBot/Modules/Core/Enums/CoreFileValuesEnum.cs
@@ -1,0 +1,9 @@
+﻿namespace AzzyBot.Modules.Core.Enums;
+
+internal enum CoreFileValuesEnum
+{
+    Commit,
+    CompileDate,
+    LoC,
+    None
+}